Pridanie ApiResponse, EmailAlreadyExistsException, GlobalExceptionHandler, ObjectNotFoundException. Uprava entity Employee, EmployeeRestController a EmployeeRepository

+++ src/main/java/sk/ukf/uloha2/service/EmployeeServiceImpl.java
@@ -6,6 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import sk.ukf.uloha2.dao.EmployeeDAO;
 import sk.ukf.uloha2.dao.EmployeeRepository;
 import sk.ukf.uloha2.entity.Employee;
+import sk.ukf.uloha2.exception.EmailAlreadyExistsException;
+import sk.ukf.uloha2.exception.ObjectNotFoundException;
 
 import java.util.List;
 
@@ -54,18 +56,33 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public Employee findById(int id) {
-        return this.employeeRepository.findById(id).orElse(null);
+        return this.employeeRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Employee",id));
     }
 
     @Transactional
     @Override
     public Employee save(Employee employee) {
+        if (employee.getId() == 0) {
+            if(employeeRepository.existsByEmail(employee.getEmail())) {
+                throw new EmailAlreadyExistsException(employee.getEmail());
+            }
+        } else{
+            Employee existingWithEmail = employeeRepository.findByEmail(employee.getEmail()).orElse(null);
+
+            if (existingWithEmail != null && existingWithEmail.getId() != employee.getId()) {
+                throw new EmailAlreadyExistsException(employee.getEmail());
+            }
+        }
+
         return this.employeeRepository.save(employee);
     }
 
     @Transactional
     @Override
     public void deleteById(int id) {
+        if(!employeeRepository.existsById(id)) {
+            throw new ObjectNotFoundException("Employee",id);
+        }
         this.employeeRepository.deleteById(id);
     }
 }
